홈 슬라이드 수정
