+++ src/pages/Slide.css
@@ -1,0 +1,23 @@
+.slick-prev,
+.slick-next {
+  position: absolute !important;
+  z-index: 99999 !important;
+  width: 40px;
+  height: 40px;
+}
+
+.slick-prev {
+  left: 0;
+  top: 38%;
+}
+
+.slick-next {
+  right: 0;
+  top: 38%;
+}
+
+.slick-prev:before,
+.slick-next:before {
+  color: black !important;
+  font-size: 2rem;
+}
